Each attempt to persist/publish an event stream will "re-version" the events to be committed.
Stamping the events to be committed with the appropriate aggregate version prior to save ensures that they
will always be associated with the appropriate aggregate version.  This is most apparent when an optimistic
concurrency exception occurs in which no business-level conflicts occur.  The events are then committed.
If we do not stamp the events with the appropriate version, the events themselves would be persisted
with the old (and now incorrect) version number.